feat(logging): RPE-first set logging with quick-select picker
Replace buried RPE text field with prominent horizontal chip picker
(6–10, 0.5 steps) on both strength and cardio set input forms.
Color-code completed set badges: green 6-7.5, amber 8-8.5, orange
9-9.5, red 10. Badge shown in active logging tiles, workout history
set rows, and session summary screen using shared RpeBadge widget.

Change rpe field type from int to double/Float across the full stack
(Prisma, Zod validation, Drift table, DTOs, domain model, repository
interfaces) to support half-increment values. PostgreSQL migration
alters the column type; Drift schema v3 uses TableMigration to
recreate set_logs with the updated column definition.

Fixes: warmup flag not reset between sets, multi-dot decimal input
silently dropping weight data, sync DTO truncating 0.5 RPE values
on pull-sync, rpe server validation floor raised from 1 to 6.

+++ app/lib/features/active_session/presentation/screens/active_workout_screen.dart
@@ -60,7 +60,7 @@ class _ActiveWorkoutBodyState extends ConsumerState<_ActiveWorkoutBody> {
   Future<void> _logSet({
     int? reps,
     double? weightKg,
-    int? rpe,
+    double? rpe,
     String? tempo,
     bool isWarmup = false,
     int? durationSec,
@@ -360,7 +360,7 @@ class _ExerciseLogger extends StatelessWidget {
   final void Function({
     int? reps,
     double? weightKg,
-    int? rpe,
+    double? rpe,
     String? tempo,
     bool isWarmup,
     int? durationSec,
@@ -477,7 +477,7 @@ class _ExerciseLogger extends StatelessWidget {
                       int? durationSec,
                       double? distanceM,
                       int? heartRate,
-                      int? rpe,
+                      double? rpe,
                     }) =>
                         onLogSet(
                       durationSec: durationSec,
@@ -495,7 +495,7 @@ class _ExerciseLogger extends StatelessWidget {
                     onLog: ({
                       int? reps,
                       double? weightKg,
-                      int? rpe,
+                      double? rpe,
                       String? tempo,
                       bool isWarmup = false,
                     }) =>
